+++ CHANGES.md
@@ -1,6 +1,18 @@
 # 変更履歴
 
+- UPDATE
+    - 下位互換がある変更
+- ADD
+    - 下位互換がある追加
+- CHANGE
+    - 下位互換のない変更
+- FIX
+    - バグ修正
+
 ## develop
+
+- [UPDATE] WebRTC のバージョンを M79@{#5} に上げた
+- [UPDATE] Windows 版の WebRTC ライブラリを shiguredo-webrtc-build からダウンロードするようにした
 
 ## 1.0.1